// src-backup/types/service.ts
export interface Service {

    /*
    |--------------------------------------------------------------------------
    | Primary
    |--------------------------------------------------------------------------
    */

    id?: string;

    client_id: string;

    /*
    |--------------------------------------------------------------------------
    | Service Information
    |--------------------------------------------------------------------------
    */

    name: string;

    description?: string;

    category?: string;

    /*
    |--------------------------------------------------------------------------
    | Booking
    |--------------------------------------------------------------------------
    */

    duration_minutes: number;

    price: number;

    currency: string;

    /*
    |--------------------------------------------------------------------------
    | Appearance
    |--------------------------------------------------------------------------
    */

    color?: string;

    /*
    |--------------------------------------------------------------------------
    | Settings
    |--------------------------------------------------------------------------
    */

    online_booking: boolean;

    is_active: boolean;

    /*
    |--------------------------------------------------------------------------
    | Timestamps
    |--------------------------------------------------------------------------
    */

    created_at?: string;

    updated_at?: string;

}